simplify the export
