+++ packages/core/src/Nodes/NodeInstance.ts
@@ -2,13 +2,7 @@ import { Engine } from '../Execution/Engine';
 import { Fiber } from '../Execution/Fiber';
 import { IGraphApi } from '../Graphs/Graph';
 import { Socket } from '../Sockets/Socket';
-import { Node, NodeConfiguration } from './Node';
-import {
-  IAsyncNodeDefinition,
-  IEventNodeDefinition,
-  IFlowNodeDefinition,
-  IFunctionNodeDefinition
-} from './NodeDefinitions';
+import { NodeConfiguration } from './Node';
 import { readInputFromSockets, writeOutputsToSocket } from './NodeSockets';
 import { INodeDescription } from './Registry/NodeDescription';
 
@@ -93,150 +87,3 @@ export const makeNodeInstance = (node: INode) => {
     writeOutput
   };
 };
-
-export class FlowNodeInstance<TFlowNodeDefinition extends IFlowNodeDefinition>
-  extends Node<NodeType.Flow>
-  implements IFlowNode
-{
-  private triggeredInner: TFlowNodeDefinition['triggered'];
-  private state: TFlowNodeDefinition['initialState'];
-  private readonly outputSocketKeys: string[];
-
-  constructor(
-    nodeProps: Omit<INode, 'nodeType'> &
-      Pick<TFlowNodeDefinition, 'triggered' | 'initialState'>
-  ) {
-    super({ ...nodeProps, nodeType: NodeType.Flow });
-    this.triggeredInner = nodeProps.triggered;
-    this.state = nodeProps.initialState;
-    this.outputSocketKeys = nodeProps.outputs.map((s) => s.name);
-  }
-
-  public triggered(fiber: Fiber, triggeringSocketName: string) {
-    this.state = this.triggeredInner({
-      commit: (outFlowName, fiberCompletedListener) =>
-        fiber.commit(this, outFlowName, fiberCompletedListener),
-      read: this.readInput,
-      write: this.writeOutput,
-      state: this.state,
-      outputSocketKeys: this.outputSocketKeys,
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      triggeringSocketName
-    });
-  }
-}
-
-export class AsyncNodeInstance<TAsyncNodeDef extends IAsyncNodeDefinition>
-  extends Node<NodeType.Async>
-  implements IAsyncNode
-{
-  private triggeredInner: TAsyncNodeDef['triggered'];
-  private disposeInner: TAsyncNodeDef['dispose'];
-  private state: TAsyncNodeDef['initialState'];
-
-  constructor(
-    node: Omit<INode, 'nodeType'> &
-      Pick<TAsyncNodeDef, 'triggered' | 'initialState' | 'dispose'>
-  ) {
-    super({ ...node, nodeType: NodeType.Async });
-
-    this.triggeredInner = node.triggered;
-    this.disposeInner = node.dispose;
-    this.state = node.initialState;
-  }
-
-  triggered = (
-    engine: Pick<Engine, 'commitToNewFiber'>,
-    triggeringSocketName: string,
-    finished: () => void
-  ) => {
-    this.triggeredInner({
-      read: this.readInput,
-      write: this.writeOutput,
-      commit: (outFlowname, fiberCompletedListener) =>
-        engine.commitToNewFiber(this, outFlowname, fiberCompletedListener),
-      configuration: this.configuration,
-      graph: this.graph,
-      finished,
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      triggeringSocketName
-    });
-  };
-  dispose = () => {
-    this.state = this.disposeInner({ state: this.state });
-  };
-}
-
-export class EventNodeInstance<TEventNodeDef extends IEventNodeDefinition>
-  extends Node<NodeType.Event>
-  implements IEventNode
-{
-  private initInner: TEventNodeDef['init'];
-  private disposeInner: TEventNodeDef['dispose'];
-  private state: TEventNodeDef['initialState'];
-  private readonly outputSocketKeys: string[];
-
-  constructor(
-    nodeProps: Omit<INode, 'nodeType'> &
-      Pick<TEventNodeDef, 'init' | 'dispose' | 'initialState'>
-  ) {
-    super({ ...nodeProps, nodeType: NodeType.Event });
-    this.initInner = nodeProps.init;
-    this.disposeInner = nodeProps.dispose;
-    this.state = nodeProps.initialState;
-    this.outputSocketKeys = nodeProps.outputs.map((s) => s.name);
-  }
-
-  init(engine: Engine): any {
-    this.state = this.initInner({
-      read: this.readInput,
-      write: this.writeOutput,
-      state: this.state,
-      outputSocketKeys: this.outputSocketKeys,
-      commit: (outFlowname, fiberCompletedListener) =>
-        engine.commitToNewFiber(this, outFlowname, fiberCompletedListener),
-      configuration: this.configuration,
-      graph: this.graph
-    });
-  }
-
-  dispose(): void {
-    this.disposeInner({
-      state: this.state
-    });
-  }
-}
-
-export class FunctionNodeInstance<
-    TFunctionNodeDef extends IFunctionNodeDefinition
-  >
-  extends Node<NodeType.Function>
-  implements IFunctionNode
-{
-  private execInner: TFunctionNodeDef['exec'];
-  constructor(
-    nodeProps: Omit<INode, 'nodeType'> & Pick<TFunctionNodeDef, 'exec'>
-  ) {
-    super({ ...nodeProps, nodeType: NodeType.Function });
-
-    this.execInner = nodeProps.exec;
-  }
-
-  exec(node: INode) {
-    this.execInner({
-      read: (name) =>
-        readInputFromSockets(node.inputs, name, node.description.typeName),
-      write: (name, value) =>
-        writeOutputsToSocket(
-          node.outputs,
-          name,
-          value,
-          node.description.typeName
-        ),
-      configuration: this.configuration,
-      graph: this.graph
-    });
-  }
-}
